Revert "Feature/ib 07"
