+++ Moment/Shared/ContentView.swift
@@ -1,0 +1,20 @@
+//
+//  ContentView.swift
+//  Shared
+//
+//  Created by Beatriz Carlos on 14/05/21.
+//
+
+import SwiftUI
+
+struct ContentView: View {
+    var body: some View {
+        Text("Hello, world!").padding()
+    }
+}
+
+struct ContentView_Previews: PreviewProvider {
+    static var previews: some View {
+        ContentView()
+    }
+}
