--- Moment/Shared/ViewController/ViewController.swift
@@ -1,8 +1,0 @@
-//
-//  ViewController.swift
-//  Moment
-//
-//  Created by Beatriz Carlos on 18/05/21.
-//
-
-import Foundation
