--- Moment/Shared/ViewModel/ViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  ViewModel.swift
-//  Moment
-//
-//  Created by Beatriz Carlos on 18/05/21.
-//
-
-import Foundation
