--- Moment/macOS/MomentAppMacOS.swift
@@ -1,17 +1,0 @@
-//
-//  MomentAppMacOS.swift
-//  Moment
-//
-//  Created by Beatriz Carlos on 18/05/21.
-//
-
-import SwiftUI
-
-@main
-struct MomentApp: App {
-    var body: some Scene {
-        WindowGroup {
-            ContentView().frame(width: 440, height: 600, alignment: .center)
-        }
-    }
-}
